+++ lib/src/auth/auth.dart
@@ -10,6 +10,12 @@ class AuthorizationManager {
     return _instance;
   }
 
+  static Future<String> getPerferenceType() async {
+    final SharedPreferences preferences = await SharedPreferences.getInstance();
+    final String type = preferences.getString('authnull-preference-type');
+    return type;
+  }
+
   AuthorizationManager._();
 
   bool _inited = false;